Changes features to feature collections
Each document in the collections is now a feature collection and contains GeoJSON arrays of Features, so that a FeatureCollection can be rendered from the document. This allows me to represent multiple geometries for each feature with different style options, hopefully allowing me to group more feature documents together without affecting the map itself.

Also builds on this to group the GUTE stuff together.

It will probably be feasible to remove some duplication of rendering code from map.js and feature.html now, and there’s some Liquid de-duplication that can be done in map.js